requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Werkzeug>=2.3.0
pandas>=1.5.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0
reportlab>=4.0.0
joblib>=1.2.0
openpyxl>=3.1.0
xlsxwriter>=3.0.0
xgboost>=2.0.0
